requirements-bot.txt
python-telegram-bot==21.5
google-cloud-storage==2.18.2
google-auth==2.35.0
httpx==0.27.2
